deploymlapp/requirements.txt
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
future==0.18.2
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
opencv-python==4.2.0.34
Pillow==8.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
skorch==0.9.0
tabulate==0.8.7
threadpoolctl==2.1.0
tifffile==2020.10.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
visitor==0.1.3
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.3
